lambdas/model_predict/requirements.txt
scikit-learn
pandas
numpy
psycopg2-binary
python-dotenv
boto3